Setup CropView to hold CropScrollView and CropOverlayView
